% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{biodome}
\alias{biodome-package}
\alias{biodome}
\title{Biodome}
\description{
This is still the pre-alpha version. Notes about found bugs and suggestions are more than welcome!
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adamkocsis/biodome}
}

}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
