% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bradypus}
\alias{bradypus}
\title{Occurrences of Bradypus variegatus from GBIF}
\format{
A \code{data.frame} with 1466 observations and 6 variables:
	\describe{
		\item{\code{scientificName}}{Taxon name.}
		\item{\code{basisOfRecord}}{The basis of the record.}
		\item{\code{elevation}}{Elevation of recors}
		\item{\code{decimalLongitude}}{Longitude in decimal degrees.}
		\item{\code{decimalLatitude}}{Latitude in decimal degrees.}
	}
}
\usage{
data(bradypus)
}
\description{
This is an example occurrence record \code{data.frame} of a single terrestrial species.
GBIF.org (30 June 2026) GBIF Occurrence Download https://doi.org/10.15468/dl.s7zhvp
}
\keyword{datasets}
