% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{cenrad}
\alias{cenrad}
\title{Calculate range with the centroid-radius method}
\usage{
cenrad(
  coordMat,
  centroid = centroid_points(coordMat),
  plot = FALSE,
  plot.args = NULL,
  q = 0.95
)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{centroid}{The centroid given with a latitude-longitude coordinate pair.}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}

\item{q}{The quantile that will be returned as an estimate}
}
\value{
A list with an estimate an two indices the rows of the input matrix that represent the longest great circle (or one of them).
}
\description{
Calculate range with the centroid-radius method
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=5, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
centDist <- cenrad(coordMat, plot=TRUE)

lines(rbind(centDist$centroid,
		coordMat[which(centDist$estimate==centDist$distances)[1],]),
		col="blue", lwd=3)
}
