% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroid.R
\name{centroid}
\alias{centroid}
\alias{centroid,matrix-method}
\alias{centroid,data.frame-method}
\title{Calculate centroid of a point cloud}
\usage{
centroid(x, ...)

\S4method{centroid}{matrix}(
  x,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL
)

\S4method{centroid}{data.frame}(
  x,
  tax = NULL,
  long = "long",
  lat = "lat",
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL
)
}
\arguments{
\item{x}{Eiher a 2-column numeric matrix with two columns: longitudes and latitudes, or a \code{data.frame} with these columns.}

\item{...}{Additional arguments passed to class-specific methods.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE}).}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}), if here is any.}

\item{plot.args}{List arguments passed to the plotting function: \code{points}.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}
}
\value{
Either a single numeric or a list with an estimate and other information.
}
\description{
Calculate centroid of a point cloud
}
\examples{
# Records
data(pinna)
nobilis <- pinna[pinna$species=="Pinna nobilis", ]
plot(nobilis[c("decimalLongitude", "decimalLatitude")], pch=16, col="#00BBAA66")

# Number of unique coordinate pairs
cent <- centroid(nobilis, long="decimalLongitude", lat="decimalLatitude")

points(cent[1], cent[2], col="darkred", pch=3, lwd=4, cex=4)
}
