% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\name{circleOcc}
\alias{circleOcc}
\title{Circular occurrence model}
\usage{
circleOcc(g, p, s = NULL, centers = NULL)
}
\arguments{
\item{g}{A trigrid object defined in icosa}

\item{p}{Target occupancy expressed as a proportion. Either a single value, or a vector.}

\item{s}{Single integer, the number of taxa - only if p is a single probability value.}
}
\description{
Circular occurrence model
}
\examples{

# Single probabilitty, repeated s number of times
repeated <- uniformOcc(p=0.3, s=100, k=40)
}
