% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_circles.R
\name{circumcenter}
\alias{circumcenter}
\title{Identify the center of a small circle based on three points of a sphere}
\usage{
circumcenter(x, origin = c(0, 0, 0), output = "polar")
}
\arguments{
\item{x}{A matrix of 3 points, either polar longitude and latitude coordinates, or XYZ Cartesian coordinates.}

\item{origin}{The origin of the sphere.}

\item{output}{If set to "polar" then the function will return the longitude and latitude of the small circle center.}
}
\value{
A list with two elements, the center o the small circle and the surface radius of the small circle.
}
\description{
Identify the center of a small circle based on three points of a sphere
}
\examples{
# generate 3 points on a sphere
ps <- icosa::rpsphere(3, output="polar")
small <- circumcenter(x=ps)
circle<- smallcircles(x=small$center, r=small$r, output="polar")

plot(NULL, NULL, xlim=c(-180, 180), ylim=c(-90, 90))
points(ps, col="red", pch=3, cex=4)
points(small$center, col="blue", pch=16)
points(circle, col="green", pch=16)
}
