% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gappiness.R
\name{gappiness}
\alias{gappiness}
\alias{gappiness,character,trigrid-method}
\alias{gappiness,matrix,trigrid-method}
\title{The gappiness a of shape}
\usage{
gappiness(x, s, ...)

\S4method{gappiness}{character,trigrid}(x, s, exclude = NULL, full = FALSE)

\S4method{gappiness}{matrix,trigrid}(
  x,
  s,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  exclude = NULL
)
}
\arguments{
\item{x}{The list of faces that are part of the shape.}

\item{s}{Spatial discretization strucutre (currently only \code{trigrid}-class icosahedral grid).}

\item{...}{Arguments passed to class-specific methods.}

\item{exclude}{The list of faces that is to be excluded from the calculation}

\item{full}{\code{logical}, should only the estimate (\code{FALSE}) be returned, or additional data as well?(\code{TRUE}).}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE})}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}), if here is any.}

\item{plot.args}{List arguments passed to the plotting function: \code{plot}.}
}
\value{
A proportion that gives the ratio of hole-cells compared to all occupied cells.
}
\description{
Proportion of gap cells in an icosahedral grid.
}
\details{
Gappiness refers to proporion of area that are covered by the internal gaps that defined by a set of discretized cells or a point set that covers some cells in a discretization structre.
}
\examples{
# 1. Only cells
# create a grid
hex <- hexagrid(2, sf=TRUE)

# an example shape
shape <- paste0("F", c(4, 5, 11, 13, 15, 21, 24, 26, 32, 33, 34, 35, 36))
plot(hex, border="gray80")
plot(hex, shape, col="red", add=TRUE)

# the gappiness
gap <- gappiness(shape, hex, full=TRUE)
gap
# plot the first gap
plot(hex, names(gap$holes[gap$holes==1]), col="#00555577", add=TRUE)

# points
set.seed(7)
rand <- icosa::rpsphere(20, output="polar")
occ <- locate(hex, rand)
plot(hex, border="gray80")
plot(hex, occ, col="red", add=TRUE)
points(rand, col="green", pch=3, cex=2)

# gappiness
gap <- gappiness(occ, hex, full=TRUE)
plot(hex, names(gap$holes), col="orange", add=TRUE)
}
