% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latrange.R
\name{latrange}
\alias{latrange}
\alias{latrange,matrix-method}
\alias{latrange,data.frame-method}
\title{Calculate latitudinal ranges}
\usage{
latrange(x, ...)

\S4method{latrange}{matrix}(
  x,
  long = NULL,
  lat = NULL,
  q = 1,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE
)

\S4method{latrange}{data.frame}(
  x,
  tax = NULL,
  q = 1,
  long = "long",
  lat = "lat",
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE
)
}
\arguments{
\item{x}{Either a 2D numeric \code{matrix} with two columns: longitudes and latitudes, a \code{data.frame} with the same information.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{q}{\code{numeric}, a value between 0 and 1, the quantile.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE})}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}), if here is any.}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}
}
\value{
Either a single numeric or a list with an estimate and other information. If iterated using `tax`, then either a named vector or list of lists.
}
\description{
Calculate latitudinal ranges
}
\examples{
# 1. Records
data(pinna)
plot(pinna[c("decimallongitude", "decimallatitude")], pch=16, col="#00BBAA66")

# Number of unique coordinate pairs
lr <- latrange(pinna, long="decimallongitude", lat="decimallatitude", full=TRUE)

abline(h=lr$range, col="darkred", lty=3, lwd=4)
}
