% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdist.R
\name{maxdist}
\alias{maxdist}
\alias{mgcd}
\alias{maxdist,matrix,missing-method}
\alias{maxdist,data.frame,missing-method}
\title{Calculate ranges with the maximum distance method}
\usage{
maxdist(x, s, ...)

mgcd(x, ...)

\S4method{maxdist}{matrix,missing}(
  x,
  dm = NULL,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  q = 1
)

\S4method{maxdist}{data.frame,missing}(
  x,
  long = "long",
  lat = "lat",
  tax = NULL,
  dm = NULL,
  duplicates = FALSE,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  listout = FALSE
)
}
\arguments{
\item{x}{Either a 2D numeric \code{matrix} with two columns: longitudes and latitudes, a \code{data.frame} with the same information.}

\item{s}{Substitute spatial structure. An icosahedral grid object the inherits from the \code{trigrid} class. Providing this argumnet reduces the point cloud to the centers of the grid cells. (not yet)}

\item{...}{Additional arguments passed to class-specific methods.}

\item{dm}{If there is a pre-made distance matrix, it can be plugged in here. If this is provided, the default coordinates will not be used.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE})}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}

\item{full}{\code{logical}, should only the estimate (\code{FALSE}) be returned, or additional data as well?(\code{TRUE}).}

\item{q}{\code{numeric}, a value between 0 and 1, the quantile.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}
}
\value{
A list with an estimate an two indices the rows of the input matrix that represent the longest great circle (or one of them).
}
\description{
This family of metrics rely on the maximum distance within a point cloud
}
\details{
This metrics includes maximum great circle distance and similar methods.
}
\examples{
# 1. Records
data(pinna)
# Subset to Pinna nobilis
nobilis <- pinna[pinna$species=="Pinna nobilis", ]
plot(nobilis[c("decimalLongitude", "decimalLatitude")], pch=16, col="#00BBAA66")

# 2. calculate and visualize
mgcd <- maxdist(nobilis, long="decimalLongitude", lat="decimalLatitude", plot=TRUE, full=TRUE)

}
