% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstlength.R
\name{mstlength}
\alias{mstlength}
\alias{mstlength,matrix,missing-method}
\alias{mstlength,data.frame,missing-method}
\title{Calculate ranges with the minimum spanning tree length method}
\usage{
mstlength(x, s, ...)

\S4method{mstlength}{matrix,missing}(
  x,
  dm = NULL,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  q = 1
)

\S4method{mstlength}{data.frame,missing}(
  x,
  long = "long",
  lat = "lat",
  tax = NULL,
  dm = NULL,
  duplicates = FALSE,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE
)
}
\arguments{
\item{x}{Either a 2D numeric \code{matrix} with two columns: longitudes and latitudes, a \code{data.frame} with the same information.}

\item{s}{Structure to replace the points, either missing (coordinate pairs) or a \code{trigrid} (icosahedral grid from the package icosa).}

\item{...}{Additional arguments passed to class-specific methods.}

\item{dm}{If there is a pre-made distance matrix, it can be plugged in here. If this is provided, the default coordinates will not be used.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE})}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}

\item{full}{\code{logical}, should only the estimate (\code{FALSE}) be returned, or additional data as well?(\code{TRUE}).}

\item{q}{\code{numeric}, a value between 0 and 1, the quantile.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}
}
\value{
A list with an estimate an two indices the rows of the input matrix that represent the length of the tree (or one of them).
}
\description{
This family of metrics rely on constructing a minimum spanning tree from the distances between the points, and use its length to describe spreading.
}
\details{
This metrics includes maximum great circle distance and similar methods.
}
\examples{
# 1. Records
data(pinna)
# Subset to Pinna nobilis
nobilis <- pinna[pinna$species=="Pinna nobilis", ]
plot(nobilis[c("decimalLongitude", "decimalLatitude")], pch=16, col="#00BBAA66")

# 2. calculate and visualize
mst <- mstlength(nobilis, long="decimalLongitude", lat="decimalLatitude", plot=TRUE, full=TRUE)
mst$estimate
}
