% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{orange}
\alias{orange-package}
\alias{orange}
\title{Orange}
\description{
Spherical Ranges and Parametrization of Geographic Shapes
}
\details{
A collation of tools and metrics to characterize distribution data on the surface of a sphere or an ellipsoid. The primary group of these metrics is those that describe the extent of a distribution (geographic ranges). The calculation of geographic ranges can be executed using point coordinates data, vector polygons, as well as cells on a discretized sphere. Besides ensuring the use a geometrically correct implementations, the package offers the exploration of partial results for visual diagnostics.
This is still the pre-alpha version. Notes about found bugs and suggestions are more than welcome!
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adamkocsis/orange/issues}
}

}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
