% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinna}
\alias{pinna}
\title{Occurrences of the genus Pinna from OBIS}
\format{
A \code{data.frame} with 1466 observations and 6 variables:
	\describe{
		\item{\code{scientificName}}{Taxon name.}
		\item{\code{species}}{Species name.}
		\item{\code{basisOfRecord}}{The basis of the record.}
		\item{\code{year}}{Year.}
		\item{\code{depth}}{Depth in meters.}
		\item{\code{decimalLongitude}}{Longitude in decimal degrees.}
		\item{\code{decimalLatitude}}{Latitude in decimal degrees.}
	}
}
\usage{
data(pinna)
}
\description{
This is an example occurrence record \code{data.frame} with multiple taxa. Occurrence records downloaded from OBIS on 2026-07-01.
}
\keyword{datasets}
