% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shull_plots.R
\name{plot.shull}
\alias{plot.shull}
\title{Plotting of spherical hulls}
\usage{
\method{plot}{shull}(x, add = FALSE, ...)
}
\arguments{
\item{x}{Spherical hull object.}

\item{add}{Logical parameter specifying whether the previous plot should be overwritten.}

\item{...}{Additional plotting parameters as defined in par.}
}
\value{
The function has no return value.
}
\description{
Plotting of spherical hulls
}
