% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{presences}
\alias{presences}
\title{Generate presences based on different models}
\usage{
presences(n, method = "kent", ...)
}
\arguments{
\item{n}{The number of presences ot generate}

\item{method}{The method to generate presences}
}
\description{
Generate presences based on different models
}
