% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{presences_kent}
\alias{presences_kent}
\title{Generate random presences based on the Kent distribution}
\usage{
presences_kent(n, kappa, centers = NULL, beta = 0, listout = FALSE)
}
\arguments{
\item{n}{Either a single integer or a vector with the number of points to be generated in one distribution.}

\item{kappa}{A vector of kappa parameters of the Kent distribution. kappa = a indicates a uniform distribution, the higher kappa is, the concentrated the presences.}

\item{centers}{The center points around which the points are to created (matrix). If \code{NULL}, then these will be randomized.}

\item{beta}{The ovalness parameter of the Kent distribution}

\item{listout}{For multiple distributions, should the outputs be separated (list) or returned in a single matrix?}
}
\value{
A two-column matrix of point coordinates or a list of matrices.
}
\description{
The function is a wrapper around the \code{rkent} function in the \code{Directional} package.
}
\examples{
# uniform points on a sphere
uniform <- presences_kent(n=400, kappa=0)
# points around the 50 lon, 80 lat point
conc <- presences_kent(n=100, kappa=50, centers=matrix(c(50,80), ncol=2))
plot(NULL, NULL, xlim=c(-180, 180), ylim=c(-90,90))
points(uniform, col="red", pch=1)
points(conc, col="blue", pch=3)
}
