% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{presences_kent_icosa}
\alias{presences_kent_icosa}
\title{Generate presences from an icosahedral grid frequency table using Kent distributions}
\usage{
presences_kent_icosa(x, icosa, kappa = NULL, times = 1)
}
\arguments{
\item{x}{A table or named numeric, with integers and names as the face names of an icosahedral grid}

\item{icosa}{An icosahedral grid.}

\item{kappa}{Either a single kappa value, or a vector. The argument \code{NULL} will use 40000*4 divided
by the average vertex radius of the grid (selected arbitrarily).}

\item{times}{Single integer: a multiplier indicating how many times the should the sampling be repeated.}
}
\value{
A 2-column matrix of points.
}
\description{
The function will generate presences from Kent distributions that are centered around the face centers
of the given icosahedral grid. The
}
\examples{
# example grid
hex <- hexagrid(deg=5, sf=TRUE)

# example data
sampledFaces <- paste0("F", 50:70)
tab <- rep(10, length(sampledFaces))
names(tab) <- sampledFaces

# generate points
ps <- presences_kent_icosa(x=tab, icosa=hex, kappa=NULL, times=1)

# visualize
plot(hex, tab, reset=FALSE)
points(ps, col="red", pch=3)
}
