% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_centroid_points}
\alias{ranges_centroid_points}
\title{Calculate surface centroid of coordinates}
\usage{
ranges_centroid_points(coordMat, plot = FALSE, plot.args = NULL)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{points}.}
}
\value{
A longitude and latitude value.
}
\description{
Calculate surface centroid of coordinates
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=5, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
cent <- ranges_centroid_points(coordMat, plot=TRUE)

# secondary visualization
# points(cent[1], cent[2])
}
