% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_chullplane}
\alias{ranges_chullplane}
\title{Calculate planar convex hull}
\usage{
ranges_chullplane(coordMat, gcbound = FALSE, plot = FALSE, plot.args = NULL)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{gcbound}{Logical, should the bounds of the convex hull be great circles?}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}
}
\value{
A list with an estimate an sf-object that represents the convex hull itself.
}
\description{
Only included for the sake of comparison, it usage is in general not recommended.
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=5, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
planarChull <- ranges_chullplane(coordMat, gcbound=FALSE, plot=TRUE)
# plot(planarChull$sf, col="#55000033", add=TRUE)

planarChullGC <- ranges_chullplane(coordMat, gcbound=TRUE, plot=TRUE)
# plot(planarChullGC$sf, col="#00550033", add=TRUE)
}
