% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_chullsphere}
\alias{ranges_chullsphere}
\title{Calculate spherical convex hull}
\usage{
ranges_chullsphere(coordMat, plot = FALSE, plot.args = NULL)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}
}
\value{
A list with an estimate an sf-object that represents the convex hull itself.
}
\description{
Wrapper around icosa::chullsphere. Still in development
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=5, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
sphericalChull <- ranges_chullsphere(coordMat, plot=TRUE)

# plot(sphericalChull$sf, col="#55000033", add=TRUE)
}
