% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_gappiness}
\alias{ranges_gappiness}
\alias{ranges_gappiness,character,trigrid-method}
\alias{ranges_gappiness,matrix,trigrid-method}
\title{The gappiness of shape on an icosahedral grid}
\usage{
ranges_gappiness(x, icosa, ...)

\S4method{ranges_gappiness}{character,trigrid}(x, icosa, exclude = NULL)

\S4method{ranges_gappiness}{matrix,trigrid}(x, icosa, exclude = NULL)
}
\arguments{
\item{x}{The list of faces that are part of the shape.}

\item{icosa}{The icosahedral grid.}

\item{...}{Arguments passed to class-specific methods.}

\item{exclude}{The list of faces that is to be excluded from the calculation}
}
\description{
The gappiness of shape on an icosahedral grid
}
\examples{
# create a grid
hex <- hexagrid(2, sf=TRUE)

# an example shape
shape <- paste0("F", c(4, 5, 11, 13, 15, 21, 24, 26, 32, 33, 34, 35, 36))

# the gappiness
ranges_gappiness(shape, hex)
}
