% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_mstlength}
\alias{ranges_mstlength}
\title{Calculate Minimum Spanning Tree length}
\usage{
ranges_mstlength(coordMat, dm = NULL, plot = FALSE, plot.args = NULL)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{dm}{If there is a pre-made distance matrix, it can be plugged in here.}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{lines}.}
}
\value{
A list with an estimate and the input for lines to show the MST.
}
\description{
Calculate Minimum Spanning Tree length
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=5, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
mst <- ranges_mstlength(coordMat, plot=TRUE)
# lines(mst$show)
}
