% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{ranges_occupancy}
\alias{ranges_occupancy}
\title{Calculate ranges with the occupancy method}
\usage{
ranges_occupancy(
  coordMat,
  sf = NULL,
  icosa = NULL,
  plot = FALSE,
  plot.args = NULL,
  alpha = 1
)
}
\arguments{
\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{sf}{An sf object (not yet!).}

\item{icosa}{An icosahedral grid from the package icosa.}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{lines}.}

\item{alpha}{Minimum occupancy with \code{alpha} proportion of occurrences.}
}
\value{
A list with an estimate and the input for lines to show the MST.
}
\description{
Calculate ranges with the occupancy method
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=3, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
occ <- ranges_occupancy(coordMat, icosa=hex, plot=TRUE)

# plot(hex, occ$cells, add=TRUE, col="green")
}
