% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalks.R
\name{rwdensity}
\alias{rwdensity}
\title{Random density fields defined by random walks}
\usage{
rwdensity(
  coarse,
  fine = NULL,
  sizes = NULL,
  steps = NULL,
  start = NULL,
  tabulated = TRUE
)
}
\arguments{
\item{coarse}{The coarse-resolution grid object (defines output).}

\item{fine}{The fine-resolution grid object (random walks operate on this grid).}

\item{sizes}{The final size of the shape defined in the number of cells that constitute it. If set,}

\item{steps}{The number of steps that the random walk will take, vectorized for multiple walks.  Directly controls the algorithm.}

\item{start}{The starting cell on the fine-resolution grid. Defaults to NULL (random start)}
}
\value{
A character vector, face names of \code{x}. If \code{history} is set to \code{FALSE} it includes every face name only once, in the order of visitation. If \code{history=TRUE} face names can be duplicated following the history of the random walk.
}
\description{
Random density fields defined by random walks
}
\examples{
coarse <- hexagrid(8, sf=TRUE)
dens <- rwdensity(coarse, steps=c(200, 500))
current <- dens[dens$shape==1, "count"]
names(current) <- dens[dens$shape==1, "face"]
plot(coarse, current)

fine <- hexagrid(2)
# random shape of 15 faces
dist <- floor(rexp(15)*1500)
rand <- rwdensity(coarse, fine, steps=dist)
}
