% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalks.R
\name{rwfaces}
\alias{rwfaces}
\title{Random shapes defined with random walks}
\usage{
rwfaces(x, start = NULL, steps = NULL, size = NULL, history = FALSE)
}
\arguments{
\item{x}{The grid object}

\item{start}{The starting cell. defaults to NULL (random start)}

\item{steps}{The number of steps that the random walk will take. Directly controls the algorithm.}

\item{size}{The final size of the shape defined in the number of cells that constitute it. If set,}

\item{history}{Logical flag. Should the output include the complete history of cell occurrence through the walk, or only the cells that were visited? (default)}
}
\value{
A character vector, face names of \code{x}. If \code{history} is set to \code{FALSE} it includes every face name only once, in the order of visitation. If \code{history=TRUE} face names can be duplicated following the history of the random walk.
}
\description{
Random shapes defined with random walks
}
\examples{
hex <- hexagrid(8)
# random shape of 15 faces
rand <- rwfaces(hex, size=15)
}
