% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomWalks.R
\name{rwsphere}
\alias{rwsphere}
\title{Random walk on a sphere}
\usage{
rwsphere(x, step, time, output = "polar")
}
\arguments{
\item{x}{One pair of longitude and latitude coordinates in a matrix.}

\item{step}{Distance covered in a single step (km)}

\item{time}{The number of steps to take (integer)}

\item{output}{The}
}
\value{
A matrix of coordinates.
}
\description{
Random walk on a sphere
}
\examples{
randomWalk <- rwsphere(x=matrix(c(0,0), ncol=2), step=100, time=10)
}
