% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shull.R
\name{shull}
\alias{shull}
\alias{shull,matrix,missing-method}
\alias{shull,matrix,trigrid-method}
\alias{shullarea}
\alias{shullarea.default}
\alias{shullarea.shull}
\title{Spherical hull geometries and hull area calculations}
\usage{
shull(x, s, ...)

\S4method{shull}{matrix,missing}(
  x,
  method,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  sphererad = authRadius
)

\S4method{shull}{matrix,trigrid}(
  x,
  s,
  method,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  sphererad = authRadius,
  drop = TRUE
)

shullarea(x, ...)

\method{shullarea}{default}(
  x,
  method,
  metric = "area",
  full = FALSE,
  sphererad = authRadius,
  ...
)

\method{shullarea}{shull}(x, metric = "area", s = NULL, full = FALSE, ...)
}
\arguments{
\item{x}{Either a 2-column numeric matrix with two columns: longitudes and latitudes, or a \code{data.frame} with these columns.}

\item{s}{An external spatial structure to resolve the hull (e.g. a \code{hexagrid} object.)}

\item{...}{Additional arguments passed to class-specific methods.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE}).}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}), if there is any.}

\item{sphererad}{\code{numeric} The radius of the sphere used in the calculations, defaults to the authalic radius of Earth in km (6370.997).}

\item{drop}{\code{logical} In case \code{s} is provided, should the returned object be a hull-class object, or only the identifiers of \code{s} (\code{drop=FALSE})?}

\item{metric}{\code{character} What metric should be used for area of the hull? \code{area} returns the values in square of the unit of \code{sphererad} (defaults to square km),
\code{prop} returns the area as a proportion of the sphere, \code{count} returns the area as the number of components it occupies in \code{s}, when applicable.}
}
\value{
A \code{hull}-class object, a vector of identifiers (if applicable and \code{drop=FALSE}) or the area of te hull as a numeric.
}
\description{
Spherical hull geometries and hull area calculations
}
\examples{

# example data
data(kentsamples)
p <- kentsamples$dateline_m

# plotting
plot(NULL, NULL, xlim=c(-180, 180), ylim=c(-90,90), xlab="", ylab="")
points(p, pch=3, col="red")

# basic spherical hull definition (native)
cc <- shull(p, method="centroidcircle")
plot(cc, add=TRUE)

# spherical hull area calculation
shullarea(cc)

# basic hull definition (icosahedral grid)
hex <- icosa::hexagrid(spacing=8, sf=TRUE)
# dropping hull container
cc_hex <- shull(p, s=hex, method="centroidcircle")
plot(hex, cc_hex, col="#0000DD33", border="#FFFFFF77", add=TRUE)


# with shull container
cc_hex2 <- shull(p, s=hex, method="centroidcircle", drop=FALSE)
shullarea(cc_hex2, s=hex,  metric="prop")
}
