% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges.R
\name{SimpleCoordinates}
\alias{SimpleCoordinates}
\title{Function to process various coordinate input}
\usage{
SimpleCoordinates(x, long = "long", lat = "lat")
}
\arguments{
\item{x}{A data.frame with occurrence records}

\item{long}{Column name of longitudes}

\item{lat}{Column name of latitudes}
}
\value{
A matrix of longitude and latitude coordinates

A two-column numeric matrix, longitudes and latitudes.
}
\description{
Note: this will be buried in the package namespace!
}
