% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_circles.R
\name{smallcircles}
\alias{smallcircles}
\title{Generate coordinates of small circles}
\usage{
smallcircles(
  x = "random",
  r = NULL,
  r.ex = NULL,
  r.rad = NULL,
  r.deg = NULL,
  n = NULL,
  breaks = 100,
  radius = authRadius,
  origin = c(0, 0, 0),
  output = "polar",
  sf.type = "polygon",
  sf.wrap.dateline = TRUE,
  drop = TRUE
)
}
\arguments{
\item{x}{Numeric matrix or character string. Coordinates of the centers of the small circles}

\item{r}{The surface radius of the small circle expressed as great circle distance of the small circle's points from the center of the small circles.}

\item{r.ex}{The extrinsic radius of the small circles, i.e. the radius of the circle in the plane of the small circle.}

\item{r.rad}{The angle of the small circle in radians.}

\item{r.deg}{The angle of the small circle in degrees.}

\item{n}{Integer or NULL. only used if "x= random"}

\item{breaks}{Integer. The number of points to create from the small circle.}

\item{radius}{The radius of the sphere, defaults to the authalic radius of Earth.}

\item{origin}{The center of the sphere (don't touch this, unless you know what you are doing!).}

\item{output}{Output structure for the function. The value output="polar" will return the polar (longitude-latitude) coordinates of the small circles in an array.}

\item{drop}{If there is a single small circle to generate, should its array wrapper be dropped?
The setting output="cartesian" will return the 3D cartesian coordinates of the small circles in an array. The option output="sf" will return an sfc geometry collection.}
}
\value{
Either a numeric array or and sfc geometry collection.
}
\description{
Generate coordinates of small circles
}
