% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence.R
\name{uniformOcc}
\alias{uniformOcc}
\title{Uniform Bernoulli Model}
\usage{
uniformOcc(p, s = NULL, g = NULL, k = NULL, nozero = TRUE)
}
\arguments{
\item{p}{Either a single probability value, or a vector of probabilities.}

\item{s}{Single integer, the number of taxa - only if p is a single probability value.}

\item{g}{A trigrid object defined in icosa}

\item{k}{The number of spatial cells.}
}
\description{
Uniform Bernoulli Model
}
\examples{

# Single probabilitty, repeated s number of times
repeated <- uniformOcc(p=0.3, s=100, k=40)
}
