% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{emptymap}
\alias{emptymap}
\title{Empty world map (Plate Carée)}
\usage{
emptymap(
  xlim = c(-180, 180),
  ylim = c(-90, 90),
  grat = c(30, 30),
  col = "#DDDDFE",
  grat.col = "#FFFFFF",
  lwd = 2,
  thick = lwd * 2,
  border = grat.col
)
}
\arguments{
\item{xlim}{Standard \code{xlim} argument of par in degress longitude.}

\item{ylim}{Standard \code{ylim} argument of par in degrees latitude.}

\item{grat}{Graticule resolution. Use \code{grat=NULL}, if you don't want any.}

\item{col}{Plot background, RGBA possible.}

\item{grat.col}{Graticule color RGBA possible.}

\item{thick}{The 0 coordinate circles are highlighted with thicker lines, given here}

\item{border}{Border color of rectange on the outside of the plot.}
}
\value{
The function has no return value.
}
\description{
Empty world map (Plate Carée)
}
\examples{
data(kentsamples)
emptymap(l)
points(kentsamples$central_s, pch=16, col=1)
points(kentsamples$arctic_m, pch=16, col=2)
points(kentsamples$antarctic_l, pch=16, col=3)
points(kentsamples$dateline_l, pch=16, col=4)
}
