% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gappiness.R
\name{gappiness}
\alias{gappiness}
\alias{gappiness,character,trigrid-method}
\title{The gappiness a of shape}
\usage{
gappiness(x, s, ...)

\S4method{gappiness}{character,trigrid}(x, s, exclude = NULL, full = FALSE)
}
\arguments{
\item{x}{The list of faces that are part of the shape.}

\item{s}{Spatial discretization strucutre (currently only \code{trigrid}-class icosahedral grid).}

\item{...}{Arguments passed to class-specific methods.}

\item{exclude}{The list of faces that is to be excluded from the calculation}

\item{full}{\code{logical}, should only the estimate (\code{FALSE}) be returned, or additional data as well?(\code{TRUE}).}
}
\value{
A proportion that gives the ratio of hole-cells compared to all occupied cells.
}
\description{
Gappiness refers to the internal gaps that defined by a set of discretized cells or a point set that covers some cells in a discretization structre.
}
\examples{
# create a grid
hex <- hexagrid(2, sf=TRUE)

# an example shape
shape <- paste0("F", c(4, 5, 11, 13, 15, 21, 24, 26, 32, 33, 34, 35, 36))

# the gappiness
gappiness(shape, hex)
}
