% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\name{occupancy}
\alias{occupancy}
\alias{occupancy,matrix,missing-method}
\alias{occupancy,data.frame,missing-method}
\alias{occupancy,data.frame,character-method}
\alias{occupancy,matrix,trigrid-method}
\alias{occupancy,data.frame,trigrid-method}
\title{Calculate ranges with the occupancy method}
\usage{
occupancy(x, s, ...)

\S4method{occupancy}{matrix,missing}(x, long = NULL, lat = NULL, full = FALSE, prop = NULL)

\S4method{occupancy}{data.frame,missing}(
  x,
  tax = NULL,
  long = "long",
  lat = "lat",
  full = FALSE,
  listarray = TRUE,
  prop = NULL
)

\S4method{occupancy}{data.frame,character}(x, s, tax = NULL, full = FALSE, prop = NULL, listarray = TRUE)

\S4method{occupancy}{matrix,trigrid}(
  x,
  s,
  long = NULL,
  lat = NULL,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  prop = NULL
)

\S4method{occupancy}{data.frame,trigrid}(
  x,
  s,
  long = "long",
  lat = "lat",
  tax = NULL,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  prop = NULL
)
}
\arguments{
\item{x}{Eiher a 2-column numeric matrix with two columns: longitudes and latitudes, or a \code{data.frame} with these columns.}

\item{s}{Structure to be occupied, either \code{NULL} (coordinate pairs), \code{character} (column name indicating locality) or a \code{trigrid} (icosahedral grid from the package icosa).}

\item{...}{Additional arguments passed to class-specific methods.}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{full}{Logical switch indicating whether only the estimate should be shown (\code{FALSE}), or other info (i.e. the list of occupied components in \code{s}) as well.}

\item{prop}{Should counts be returned (\code{prop=NULL}), or proportions? If \code{prop="global"}, then global proportions are returned, if \code{prop="relative"}, relative proporitions are calculated.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}

\item{listarray}{If the full traceable output is required, should this be organized with list-array (native output of tapply).}

\item{q}{Minimum occupancy with \code{q} proportion of occurrences (not yet implemented!).}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{lines}.}
}
\value{
For single subsets (\code{tax=NULL}) either a single numeric or an orange list with an estimate and other information. Iterations for multiple taxa result in a named numeric vector or a list.
}
\description{
High-level, abstract function to calculate how many or what proporotion of components in a pre-specified spatial structure is occupied by a distribution dataset
}
\details{
The function by default returns counts (i.e. natural numbers) of how many discrete units are occupied.
However, there are many cases, when proportions are much more useful, which can be set with the \code{prop} argument.
Proportions are either global (\code{prop="global"}) or relative (\code{prop="relative"}). Global proportions express how much of the overall spatial structure (\code{s}) is occupied; for example, what is the proportion of cells in a grid that are occupied.
This method is only applicable, when \code{s} is defined as a spatial structure that is independent from \code{x}.
In contrast, relative proportional occupancies express proportions in the sampled set, i.e. what proportion of the overall sampled grid cells or localities are occupied by a taxon. This is particularly useful when \code{tax!=NULL}.
}
\examples{
# I. Single taxon: Pinna nobilsi
# 1. Records
data(pinna)
# Subset to Pinna nobilis
nobilis <- pinna[pinna$species=="Pinna nobilis", ]

# Number of unique coordinate pairs
cpairs <- occupancy(nobilis, long="decimalLongitude", lat="decimalLatitude")

# 2. Occupancy in icosahedral grid 
hex <- hexagrid(deg=3, sf=TRUE)
plot(nobilis[c("decimalLongitude", "decimalLatitude")], pch=16, col="#00BBAA66")
plot(hex, border="gray70", add=TRUE)

# calculate occupancy
occ <- occupancy(nobilis, s=hex, plot=TRUE, long="decimalLongitude", lat="decimalLatitude", full=TRUE)

# manual coloring from full output
plot(hex, occ$occupied, add=TRUE, col="#00BB0088")

# global proportional occupancies - relative to the grid
occprop <- occupancy(nobilis, s=hex, long="decimalLongitude", lat="decimalLatitude", prop="global")
# same as
occ$estimate/length(hex)

}
