% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy.R
\name{occupancy}
\alias{occupancy}
\alias{occupancy,data.frame,missing,missing-method}
\alias{occupancy,matrix,trigrid,missing-method}
\alias{occupancy,data.frame,trigrid,missing-method}
\title{Calculate ranges with the occupancy method}
\usage{
occupancy(x, icosa, sf, ...)

\S4method{occupancy}{data.frame,missing,missing}(x, tax, loc = NULL, long = "long", lat = "lat", duplicates = FALSE)

\S4method{occupancy}{matrix,trigrid,missing}(
  x,
  icosa,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE
)

\S4method{occupancy}{data.frame,trigrid,missing}(
  x,
  icosa,
  long = "long",
  lat = "lat",
  tax = NULL,
  duplicates = FALSE,
  q = 1,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE
)
}
\arguments{
\item{icosa}{An icosahedral grid from the package icosa.}

\item{sf}{An sf object (not yet!).}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{lines}.}

\item{coordMat}{2D numeric matrix with two columns: longitudes and latitudes.}

\item{alpha}{Minimum occupancy with \code{alpha} proportion of occurrences.}
}
\value{
A list with an estimate and the input for lines to show the MST.
}
\description{
Calculate ranges with the occupancy method
}
\examples{
# 1. Canvas
hex <- hexagrid(deg=3, sf=TRUE)
plot(hex, reset=FALSE, xlim=c(-15, 40), ylim=c(25, 63))

# 2. Records
data(pinna)

# just the coordinates
coordMat <- SimpleCoordinates(pinna, long="decimallongitude", lat="decimallatitude")
points(coordMat)

# 3. calculate and visualize
occ <- occupancy(coordMat, icosa=hex, plot=TRUE)

# plot(hex, occ$cells, add=TRUE, col="green")
}
