% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{orange}
\alias{orange-package}
\alias{orange}
\title{Orange}
\description{
This is still the pre-alpha version. Notes about found bugs and suggestions are more than welcome!
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/adamkocsis/orange}
}

}
\author{
Adam T. Kocsis (adam.t.kocsis@gmail.com)
}
