% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pinna}
\alias{pinna}
\title{Occurrences of Pinna nobilis from OBIS}
\format{
A \code{data.frame} with 1466 observations and 6 variables:
	\describe{
		\item{\code{scientificname}}{Taxon name.}
		\item{\code{basisofrecord}}{The basis of the record.}
		\item{\code{date_year}}{Year.}
		\item{\code{depth}}{Depth in meters.}
		\item{\code{decimallongiude}}{Longitude in decimal degrees.}
		\item{\code{decimallatitude}}{Latitude in decimal degrees.}
	}
}
\usage{
data(pinna)
}
\description{
Occurrence records downloaded from OBIS on 2024-10-23
}
\details{
This is an example occurrence record \code{data.frame}.
}
\keyword{datasets}
