% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radius.R
\name{radius}
\alias{radius}
\alias{radius,matrix,missing-method}
\title{Calculate ranges with the radius-group of methods}
\usage{
radius(x, s, ...)

\S4method{radius}{matrix,missing}(
  x,
  p = NULL,
  long = NULL,
  lat = NULL,
  duplicates = FALSE,
  plot = FALSE,
  plot.args = NULL,
  full = FALSE,
  q = 1
)
}
\arguments{
\item{x}{Either a 2D numeric \code{matrix} with two columns: longitudes and latitudes, a \code{data.frame} with the same information.}

\item{s}{Structure to replace the points, either missing (coordinate pairs) or a \code{trigrid} (icosahedral grid from the package icosa).}

\item{long}{\code{character}, column name of the longitudes.}

\item{lat}{\code{character}, column name of the latitudes.}

\item{duplicates}{\code{logical}, should identical coordinates be included in the calculation (default is \code{FALSE})}

\item{plot}{Logical, should the result be plotted? Will plot over active plot (as in \code{add=TRUE}).}

\item{plot.args}{List arguments passed to the plotting function: \code{sf::plot}.}

\item{full}{\code{logical}, should only the estimate (\code{FALSE}) be returned, or additional data as well?(\code{TRUE}).}

\item{q}{\code{numeric}, a value between 0 and 1, the quantile.}

\item{tax}{\code{character}, used only in the \code{data.frame} method. Column name of groups (e.g. taxa) that allows the iteration of the method for multiple groups.}

\item{dm}{If there is a pre-made distance matrix, it can be plugged in here. If this is provided, the default coordinates will not be used.}
}
\value{
A list with an estimate an two indices the rows of the input matrix that represent the length of the tree (or one of them).
}
\description{
This family of metrics rely on estimating the distance of an extent from a fixed point.
}
\details{
This group of methods rely on calculating single distances.
}
\examples{
# 1. Records
data(pinna)

# 2. calculate and visualize
rad <- radius(pinna, long="decimallongitude", lat="decimallatitude", plot=TRUE, full=TRUE)

}
